Move SMW / Bucket / Cargo tools into pack directories
Relocate the six extension tool source files and their tests under
`src/tools/extensions/<id>/` so each extension's code lives together
with its pack module. Pure file moves with import-path adjustments;
no logic changes.

`git log --follow <new-path>` recovers the pre-move history.
