+++ src/tools/extensions/bucket/index.ts
@@ -1,5 +1,5 @@
 import type { ExtensionPack } from '../types.js';
-import { bucketQuery } from '../../bucket-query.js';
+import { bucketQuery } from './bucket-query.js';
 
 export const bucketPack: ExtensionPack = {
 	id: 'bucket',
