+++ src/tools/extensions/cargo/cargo-describe-table.ts
@@ -1,7 +1,7 @@
 import { z } from 'zod';
 import type { CallToolResult, ToolAnnotations } from '@modelcontextprotocol/sdk/types.js';
-import type { Tool } from '../runtime/tool.js';
-import type { ToolContext } from '../runtime/context.js';
+import type { Tool } from '../../../runtime/tool.js';
+import type { ToolContext } from '../../../runtime/context.js';
 
 const inputSchema = {
 	table: z.string().min(1).describe('Cargo table name. Use cargo-list-tables to discover.'),
