+++ src/tools/extensions/smw/index.ts
@@ -1,6 +1,6 @@
 import type { ExtensionPack } from '../types.js';
-import { smwQuery } from '../../smw-query.js';
-import { smwListProperties } from '../../smw-list-properties.js';
+import { smwQuery } from './smw-query.js';
+import { smwListProperties } from './smw-list-properties.js';
 
 export const smwPack: ExtensionPack = {
 	id: 'smw',
