+++ tests/tools/extensions/cargo/cargo-query.test.ts
@@ -1,10 +1,13 @@
 import { describe, it, expect, vi } from 'vitest';
-import { createMockMwn } from '../helpers/mock-mwn.js';
-import { createMockMwnError } from '../helpers/mock-mwn-error.js';
-import { fakeContext } from '../helpers/fakeContext.js';
-import { cargoQuery } from '../../src/tools/cargo-query.js';
-import { dispatch } from '../../src/runtime/dispatcher.js';
-import { assertStructuredError, assertStructuredSuccess } from '../helpers/structuredResult.js';
+import { createMockMwn } from '../../../helpers/mock-mwn.js';
+import { createMockMwnError } from '../../../helpers/mock-mwn-error.js';
+import { fakeContext } from '../../../helpers/fakeContext.js';
+import { cargoQuery } from '../../../../src/tools/extensions/cargo/cargo-query.js';
+import { dispatch } from '../../../../src/runtime/dispatcher.js';
+import {
+	assertStructuredError,
+	assertStructuredSuccess,
+} from '../../../helpers/structuredResult.js';
 
 describe('cargo-query', () => {
 	it('forwards tables and optional params to action=cargoquery and unwraps rows', async () => {
